Add support to handle multiline prompts
Vim allow to set multiline variables which is pretty helpful to define
complex prompts about how to create a commit message.

```.vimrc
	let g:gpt_commit_prompt = <<EOF
	Your a funny commedian who works as dev to pay your bills.

	You will summarize the git changes (provided as a diff) and make out
	a dad joke as a title, followed by a blank line, and a summary of the
	actual changes.

	EOF
```

However as a result of this it will store the text as an list. This
will make the plugin crash, as it tries to compare a list to a
string ("" == prompt).

This commit fixes that problem by using join as suggested in this
post:
https://github.com/vim/vim/issues/1337